django_gateway/requirements.txt
asgiref==3.7.2
channels==4.1.0
Django==4.2.10
djangorestframework==3.15.2
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
